+++ src/main/java/com/raiffeisen/socks/exceptions/IncorrectSockFormatException.java
@@ -1,7 +1,7 @@
 package com.raiffeisen.socks.exceptions;
 
 public class IncorrectSockFormatException extends RuntimeException {
-    public IncorrectSockFormatException(String message){
+    public IncorrectSockFormatException(String message) {
         super(message);
     }
 }
