+++ src/main/java/com/raiffeisen/socks/exceptions/NotFoundSockException.java
@@ -1,7 +1,7 @@
 package com.raiffeisen.socks.exceptions;
 
-public class NotFoundSockException extends RuntimeException{
-    public NotFoundSockException(String message){
+public class NotFoundSockException extends RuntimeException {
+    public NotFoundSockException(String message) {
         super(message);
     }
 }
